WATER JUG PROBLEM-ARTIFICIAL INTELLIGENCE
Write a program to solve the following problem: You have two jugs, a 4-gallon and a 3-
gallon. Neither of the jugs has markings on them. There is a pump that can be used to fill the jugs with water.
How can you get exactly two gallons of water in the 4-gallon jug?